gitlab-runner: Increase the pids_limit in containers.conf
The default is usually 2k per container which we hit very
often in runners with high cpu counts. Increase it to 64k
which is arbitrary but seems to be good enough for our workloads.

There's also a cgroup limit set by systemd on the container
slices/scopes and it defaults to around 300k usually, so
64k shouldn't be an issue or ordinarily high.

Likely fix for:
- Infrastructure/Infrastructure#1226
- Infrastructure/Infrastructure#1037